Mnk 53 hook up genre calculating

+++ backend/src/main/kotlin/org/aleks616/shrendar/genre/model/GenreDto.kt
@@ -1,7 +1,6 @@
 package org.aleks616.shrendar.genre.model
 
 import java.io.Serializable
-import java.math.BigDecimal
 
 data class GenreDto(
     val id:Int?=null,
@@ -12,5 +11,5 @@ data class GenreDto(
 data class GenreDto1(
     val id:Int?=null,
     val name:String?=null,
-    var value:BigDecimal?=null
+    var value:Long?=null
 ):Serializable
